Starts refactoring of pdf player in react
