+++ plugin/pdf-player/DependencyInjection/ClarolinePdfPlayerExtension.php
@@ -11,10 +11,10 @@
 
 namespace Claroline\PdfPlayerBundle\DependencyInjection;
 
-use Symfony\Component\HttpKernel\DependencyInjection\Extension;
-use Symfony\Component\DependencyInjection\Loader\YamlFileLoader;
-use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\Config\FileLocator;
+use Symfony\Component\DependencyInjection\ContainerBuilder;
+use Symfony\Component\DependencyInjection\Loader\YamlFileLoader;
+use Symfony\Component\HttpKernel\DependencyInjection\Extension;
 
 class ClarolinePdfPlayerExtension extends Extension
 {
