+++ plugin/pdf-player/Resources/modules/pdf/Controller/PdfController.js
