+++ plugin/pdf-player/Resources/modules/pdf/Directive/ViewPdfDirective.js
