+++ plugin/pdf-player/Resources/modules/pdf/Partial/view_pdf.html
