+++ plugin/pdf-player/Resources/modules/pdf/module.js
