+++ plugin/pdf-player/Resources/modules/resources/pdf/components/resource.jsx
