+++ plugin/pdf-player/Resources/modules/resources/pdf/reducer.js
@@ -2,8 +2,8 @@ import {makeReducer} from '#/main/core/scaffolding/reducer'
 import {makeResourceReducer} from '#/main/core/resource/reducer'
 
 const reducer = makeResourceReducer({}, {
-  url: makeReducer({}, {}),
-  pdf: makeReducer({}, {})
+  pdf: makeReducer({}, {}),
+  url: makeReducer({}, {})
 })
 
 export {
