+++ plugin/pdf-player/Resources/modules/resources/pdf/selectors.js
